refactor: update operatorProfile page
