+++ src/features/operator-profile/ui/OperatorPhilosophy/OperatorPhilosophy.test.tsx
