+++ src/features/operator-profile/ui/OperatorProfileInfo/OperatorProfileInfo.test.tsx
