+++ src/features/operator-profile/ui/OperatorProfileInfo/OperatorProfileInfo.tsx
